@@ -2,9 +2,11 @@ package com.qr.core.library.rxcache;
 
 import android.content.Context;
 import android.os.Environment;
-import android.support.test.InstrumentationRegistry;
-import android.support.test.runner.AndroidJUnit4;
+import androidx.test.platform.app.InstrumentationRegistry;
+import androidx.test.ext.junit.runners.AndroidJUnit4;
 import android.util.Log;
+
+import com.qr.core.library.rxcache.keys.DynamicKey;
 
 import org.junit.Test;
 import org.junit.runner.RunWith;
@@ -13,8 +15,6 @@ import java.io.File;
 
 import io.reactivex.Observable;
 import io.reactivex.schedulers.Schedulers;
-
-import static org.junit.Assert.*;
 
 /**
  * Instrumented test, which will execute on an Android device.
