@@ -7,8 +7,7 @@ import java.util.ArrayList;
 import java.util.List;
 
 public abstract class Action {
-    private static final String PREFIX_DYNAMIC_KEY = "$d$d$d$";
-    private static final String PREFIX_DYNAMIC_GROUP_KEY = "$g$g$g$";
+    private static final String PREFIX_KEY = "$d$d";
     protected final Memory memory;
     protected final Persistence persistence;
 
@@ -17,40 +16,25 @@ public abstract class Action {
         this.persistence = persistence;
     }
 
-    protected String composeKey(String providerKey,String dynamicKey,String dynamicGroupKey){
-        return providerKey
-                + PREFIX_DYNAMIC_KEY
-                + dynamicKey
-                + PREFIX_DYNAMIC_GROUP_KEY
-                + dynamicGroupKey;
-    }
-
-    protected List<String> getKeysOnMemoryMatchingProviderKey(String providerKey){
-        List<String> keysMatchingProviderKey = new ArrayList<>();
-
-        for(String composeKeyMemory : memory.keySet()){
-            String providerKeyOnMemory = composeKeyMemory.substring(0, composeKeyMemory.lastIndexOf(PREFIX_DYNAMIC_KEY));
-            if(providerKey.equals(providerKeyOnMemory)){
-                keysMatchingProviderKey.add(providerKeyOnMemory);
-            }
+    protected String composeKey(String providerKey, Object[] dynamicKeys) {
+        StringBuilder builder = new StringBuilder(providerKey);
+        for (Object dynamicKey : dynamicKeys) {
+            builder.append(PREFIX_KEY);
+            builder.append(dynamicKey);
         }
-
-        return keysMatchingProviderKey;
+        return builder.toString();
     }
 
-    protected List<String> getKeysOnMemoryMatchingDynamicKey(String providerKey,String dynamicKey){
+    protected List<String> getKeysOnMemoryMatchingDynamicKey(String providerKey, Object[] dynamicKeys) {
         List<String> keysMatchingDynamicKey = new ArrayList<>();
-        String composeProviderKeyAndDynamicKey = providerKey + PREFIX_DYNAMIC_KEY + dynamicKey;
-        for(String composeKeyMemory : memory.keySet()){
-            String providerKeyAndDynamicKeyOnMemory = composeKeyMemory.substring(0,composeKeyMemory.lastIndexOf(PREFIX_DYNAMIC_GROUP_KEY));
-            if(composeProviderKeyAndDynamicKey.equals(providerKeyAndDynamicKeyOnMemory)){
-                keysMatchingDynamicKey.add(providerKeyAndDynamicKeyOnMemory);
+        String composeKey = composeKey(providerKey, dynamicKeys);
+
+        for (String composeKeyMemory : memory.keySet()) {
+            if (composeKeyMemory.contains(composeKey)) {
+                keysMatchingDynamicKey.add(composeKeyMemory);
             }
         }
-        return keysMatchingDynamicKey;
-    }
 
-    protected String getKeysOnMemoryMatchingDynamicGroupKey(String providerKey,String dynamicKey,String dynamicGroupKey){
-        return composeKey(providerKey,dynamicKey,dynamicGroupKey);
+        return keysMatchingDynamicKey;
     }
 }
