@@ -11,34 +11,21 @@ import javax.inject.Singleton;
 
 @Singleton
 public class EvictRecord extends Action {
+
     @Inject
     public EvictRecord(Memory memory, Persistence persistence) {
         super(memory, persistence);
     }
 
-    public void evictRecordsMatchingProviderKey(String providerKey){
-        List<String> keysOnMemoryMatchingProviderKey = getKeysOnMemoryMatchingProviderKey(providerKey);
-        for (String keyMatchingProviderKey : keysOnMemoryMatchingProviderKey) {
-            memory.evict(keyMatchingProviderKey);
-            persistence.evict(keyMatchingProviderKey);
-        }
-    }
-
-    public void evictRecordsMatchingDynamicKey(String providerKey,String dynamicKey){
-        List<String> keysOnMemoryMatchingDynamicKey = getKeysOnMemoryMatchingDynamicKey(providerKey, dynamicKey);
+    public void evictRecordsMatchingDynamicKey(String providerKey, Object[] dynamicKeys) {
+        List<String> keysOnMemoryMatchingDynamicKey = getKeysOnMemoryMatchingDynamicKey(providerKey, dynamicKeys);
         for (String keyMatchingDynamicKey : keysOnMemoryMatchingDynamicKey) {
             memory.evict(keyMatchingDynamicKey);
             persistence.evict(keyMatchingDynamicKey);
         }
     }
 
-    public void evictRecordsMatchingDynamicGroupKey(String providerKey,String dynamicKey,String dynamicGroupKey){
-        String keysOnMemoryMatchingDynamicGroupKey = getKeysOnMemoryMatchingDynamicGroupKey(providerKey, dynamicKey, dynamicGroupKey);
-        memory.evict(keysOnMemoryMatchingDynamicGroupKey);
-        persistence.evict(keysOnMemoryMatchingDynamicGroupKey);
-    }
-
-    public void evictAll(){
+    public void evictAll() {
         memory.evictAll();
         persistence.evictAll();
     }
