@@ -16,11 +16,13 @@ public class SaveRecord extends Action {
         super(memory, persistence);
     }
 
-    public <T> void save(final String providerKey,final String dynamicKey,final String dynamicGroupKey,
-              final T data,final long survivalTime){
-        String composeKey = composeKey(providerKey, dynamicKey, dynamicGroupKey);
-        Record record = new Record(data,survivalTime,System.currentTimeMillis());
-        memory.put(composeKey,record);
-        persistence.saveRecord(composeKey,record);
+    public <T> void save(final String providerKey,
+                         final Object[] dynamicKeys,
+                         final T data,
+                         final long survivalTime) {
+        String composeKey = composeKey(providerKey, dynamicKeys);
+        Record record = new Record(data, survivalTime, System.currentTimeMillis());
+        memory.put(composeKey, record);
+        persistence.saveRecord(composeKey, record);
     }
 }
