@@ -1,22 +1,19 @@
 package com.qr.core.library.rxcache.processor;
 
-import com.qr.core.library.rxcache.annotation.OnCacheStrategy;
 import com.qr.core.library.rxcache.cache.Record;
 import com.qr.core.library.rxcache.cache.TwoLayersCache;
 import com.qr.core.library.rxcache.configuration.Configure;
 import com.qr.core.library.rxcache.configuration.ConfigureProviders;
 import com.qr.core.library.rxcache.exception.CacheExpirationException;
-import com.qr.core.library.rxcache.exception.CacheNoSuchElementException;
-
+import com.qr.core.library.rxcache.keys.CacheStrategyKey;
 
 import java.lang.reflect.Method;
-import java.util.NoSuchElementException;
+import java.util.Arrays;
 import java.util.concurrent.Callable;
 
 import javax.inject.Inject;
 import javax.inject.Singleton;
 
-import io.reactivex.Maybe;
 import io.reactivex.Observable;
 import io.reactivex.ObservableSource;
 
@@ -44,49 +41,25 @@ public final class ProcessorProvidersBehaviour implements ProcessorProviders {
         twoLayersCache.evictAll();
     }
 
-    public <T> Observable<T> getCacheObservable(Configure<T> configure) {
+    private <T> Observable<T> getCacheObservable(Configure<T> configure) {
         if (configure.isEvict()) {
             // 先清掉缓存
             twoLayersCache.evictDynamicGroupKey(configure.getProviderKey(),
-                    configure.getDynamicKey(),
-                    configure.getDynamicGroupKey());
+                    configure.getDynamicKeys());
         }
 
-        if (configure.getCacheStrategy() == OnCacheStrategy.Default) {
-            // 先缓存后数据
-            Observable<T> localObservable = Maybe.fromCallable(() -> {
-                Record<T> record = twoLayersCache.retrieve(configure.getProviderKey(),
-                        configure.getDynamicKey(),
-                        configure.getDynamicGroupKey());
-                if (record != null) {
-                    return record.getData();
-                }
-
-                return null;
-            }).toObservable();
-            return Observable.concatArrayEagerDelayError(localObservable,
-                    configure.getLoaderObservable().map(t -> {
-                        twoLayersCache.save(configure.getProviderKey(),
-                                configure.getDynamicKey(),
-                                configure.getDynamicGroupKey(),
-                                t,
-                                configure.getSurvivalTime());
-                        return t;
-                    }));
-        } else if (configure.getCacheStrategy() == OnCacheStrategy.NetworkPriority) {
+        if (configure.getCacheStrategy() == CacheStrategyKey.NetworkPriority) {
             return configure.getLoaderObservable()
                     .map(t -> {
                         twoLayersCache.save(configure.getProviderKey(),
-                                configure.getDynamicKey(),
-                                configure.getDynamicGroupKey(),
+                                configure.getDynamicKeys(),
                                 t,
                                 configure.getSurvivalTime());
                         return t;
                     })
                     .onErrorResumeNext(throwable -> {
                         Record<T> record = twoLayersCache.retrieve(configure.getProviderKey(),
-                                configure.getDynamicKey(),
-                                configure.getDynamicGroupKey());
+                                configure.getDynamicKeys());
                         if (record != null) {
                             // 有缓存数据就返回缓存数据
                             return Observable.just(record.getData());
@@ -95,54 +68,45 @@ public final class ProcessorProvidersBehaviour implements ProcessorProviders {
                         // 无缓存数据 返回loader错误信息
                         return Observable.error(throwable);
                     });
-        } else if (configure.getCacheStrategy() == OnCacheStrategy.CachePriority) {
+        } else if (configure.getCacheStrategy() == CacheStrategyKey.CachePriority) {
             return Observable.fromCallable(() -> {
                 Record<T> record = twoLayersCache.retrieve(configure.getProviderKey(),
-                        configure.getDynamicKey(),
-                        configure.getDynamicGroupKey());
-                if (record == null) {
-                    throw new NoSuchElementException();
+                        configure.getDynamicKeys());
+                if (record != null) {
+                    return record.getData();
                 }
-                return record.getData();
+
+                return null;
             }).onErrorResumeNext(configure.getLoaderObservable().map(t -> {
                 twoLayersCache.save(configure.getProviderKey(),
-                        configure.getDynamicKey(),
-                        configure.getDynamicGroupKey(),
+                        configure.getDynamicKeys(),
                         t,
                         configure.getSurvivalTime());
                 return t;
             }));
-        } else if (configure.getCacheStrategy() == OnCacheStrategy.OnlyNetwork) {
+        } else if (configure.getCacheStrategy() == CacheStrategyKey.NetworkOnly) {
             return configure.getLoaderObservable().map(t -> {
                 twoLayersCache.save(configure.getProviderKey(),
-                        configure.getDynamicKey(),
-                        configure.getDynamicGroupKey(),
+                        configure.getDynamicKeys(),
                         t,
                         configure.getSurvivalTime());
                 return t;
             });
-        } else if (configure.getCacheStrategy() == OnCacheStrategy.OnlyCache) {
+        } else if (configure.getCacheStrategy() == CacheStrategyKey.CacheOnly) {
             return Observable.fromCallable(() -> {
                 Record<T> retrieve = twoLayersCache.retrieve(configure.getProviderKey(),
-                        configure.getDynamicKey(),
-                        configure.getDynamicGroupKey());
-                if (retrieve == null) {
-                    throw new CacheNoSuchElementException(String.format("RxCache: 无数据!!! ProviderKey: %s DynamicKey: %s DynamicGroupKey: %s",
-                            configure.getProviderKey(),
-                            configure.getDynamicKey(),
-                            configure.getDynamicGroupKey()));
+                        configure.getDynamicKeys());
+                if (retrieve != null) {
+                    return retrieve.getData();
                 }
-                return retrieve.getData();
+                return null;
             });
-        } else if (configure.getCacheStrategy() == OnCacheStrategy.CacheControl) {
-
+        } else if (configure.getCacheStrategy() == CacheStrategyKey.CacheControl) {
             return Observable.fromCallable(() -> {
-                Record<T> retrieve = twoLayersCache.retrieve(configure.getProviderKey(), configure.getDynamicKey(), configure.getDynamicGroupKey());
+                Record<T> retrieve = twoLayersCache.retrieve(configure.getProviderKey(),
+                        configure.getDynamicKeys());
                 if (retrieve == null) {
-                    throw new CacheNoSuchElementException(String.format("RxCache: 无数据!!! ProviderKey: %s DynamicKey: %s DynamicGroupKey: %s",
-                            configure.getProviderKey(),
-                            configure.getDynamicKey(),
-                            configure.getDynamicGroupKey()));
+                    return null;
                 }
 
                 // 重新设置过期时间 防止接口更改时间后 数据可能不过期的情况
@@ -151,18 +115,19 @@ public final class ProcessorProvidersBehaviour implements ProcessorProviders {
                 // survivalTime <= 0 为永久保存
                 if (retrieve.getSurvivalTime() > 0) {
                     if (retrieve.getPersistedTime() + retrieve.getSurvivalTime() < System.currentTimeMillis()) {
-                        throw new CacheExpirationException(String.format("RxCache: 缓存过期!!! ProviderKey: %s DynamicKey: %s DynamicGroupKey: %s",
+                        // 缓存过期 清楚缓存
+                        twoLayersCache.evictDynamicGroupKey(configure.getProviderKey(),
+                                configure.getDynamicKeys());
+                        throw new CacheExpirationException(String.format("RxCache: 缓存过期!!! ProviderKey: %s DynamicKey: %s",
                                 configure.getProviderKey(),
-                                configure.getDynamicKey(),
-                                configure.getDynamicGroupKey()));
+                                Arrays.toString(configure.getDynamicKeys())));
                     }
                 }
 
                 return retrieve.getData();
             }).onErrorResumeNext(configure.getLoaderObservable().map(t -> {
                 twoLayersCache.save(configure.getProviderKey(),
-                        configure.getDynamicKey(),
-                        configure.getDynamicGroupKey(),
+                        configure.getDynamicKeys(),
                         t,
                         configure.getSurvivalTime());
                 return t;
